Override CLI FlagEnvHinter
This commit overrides the default FlagEnvHinter variable within the [cli](https://github.com/urfave/cli) package so that it doesn't append the Flags envVar value to the end of the usage description text. This was done so that the user could be told in the usage description text that they can set an environment variable for the flag and then not have the confusion of also having the text `[$ENV]` at the end of the usage description text.

Example:

Using the default function you would get the following hint text

>--kubeconfig value  The path to the kubectl config file to use to authenticate with Kubernetes. You can also set this using the environment variable KUBECONFIG. (default: "~/.kube/config") [$KUBECONFIG]

By overriding the FlagEnvHinter the text [$KUBECONFIG] will be removed and we will get the following help text

>--kubeconfig value  The path to the kubectl config file to use to authenticate with Kubernetes. You can also set this using the environment variable KUBECONFIG. (default: \"~/.kube/config\")"

Discussion of this feature was done in the comments of the pull request https://github.com/gruntwork-io/kubergrunt/pull/59

The CLI package required an upgrade due to the FlagEnvHinter variable wasn't added till version v1.21.0 (https://github.com/urfave/cli/commit/11d45572f9727acfbc93daa8565f379d396125d6)